apps/backend/pyproject.toml
[tool.poetry]
name = "ephemeral-radar-backend"
version = "0.1.0"
description = "FastAPI backend for Ephemeral Radar"
authors = ["Radar Team <team@radarapp.io>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.0"
uvicorn = { extras = ["standard"], version = "^0.35.0" }
pydantic = "^2.11.0"
pydantic-settings = "^2.10.0"
redis = "^6.4.0"
prometheus-client = "^0.22.1"
PyJWT = "^2.10.1"
cryptography = "^46.0.2"
firebase-admin = "^6.9.0"
sentry-sdk = {extras = ["fastapi"], version = "^2.27.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.1.0"
fakeredis = "^2.31.0"
httpx = "^0.28.1"
ruff = "^0.12.0"
mypy = "^1.17.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
